requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.26.0
pandas==2.1.0
Pillow==10.0.0
opencv-python==4.8.0.76
albumentations==1.3.1
matplotlib==3.7.3
seaborn==0.12.2
scikit-learn==1.3.0
tqdm==4.66.0
tensorboard==2.14.0
jupyter==1.0.0